chore: rename `index.ts` to `nodesera.ts`
